+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 ## 0.8.3
 
 - Rename `-source_url_type` to `-content_type`
+- Bring back option to compile a CSV table from scraped data